OS X IOKit utilities refactor to allow SKIP_TABLES
