+++ osquery/events/darwin/iokit.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include <atomic>
+#include <iomanip>
 
 #include <osquery/events.h>
 #include <osquery/status.h>
@@ -18,7 +19,15 @@
 #include <CoreServices/CoreServices.h>
 #include <IOKit/IOKitLib.h>
 
+#include "osquery/core/conversions.h"
+
 namespace osquery {
+
+extern const std::string kIOUSBDeviceClassName_;
+extern const std::string kIOPCIDeviceClassName_;
+extern const std::string kIOPlatformExpertDeviceClassName_;
+extern const std::string kIOACPIPlatformDeviceClassName_;
+extern const std::string kIOPlatformDeviceClassname_;
 
 struct IOKitSubscriptionContext : public SubscriptionContext {
   std::string model_id;
@@ -45,6 +54,29 @@ struct IOKitEventContext : public EventContext {
   std::string version;
   std::string serial;
 };
+
+struct IOKitPCIProperties {
+  std::string vendor_id;
+  std::string model_id;
+  std::string pci_class;
+  std::string driver;
+
+  /// Populate IOKit PCI device properties from the "compatible" property.
+  explicit IOKitPCIProperties(const std::string& compatible);
+};
+
+std::string getIOKitProperty(const CFMutableDictionaryRef& details,
+                             const std::string& key);
+
+inline void idToHex(std::string& id) {
+  long base = 0;
+  // = AS_LITERAL(int, id);
+  if (safeStrtol(id, 10, base)) {
+    std::stringstream hex_id;
+    hex_id << std::hex << std::setw(4) << std::setfill('0') << (base & 0xFFFF);
+    id = hex_id.str();
+  }
+}
 
 using IOKitEventContextRef = std::shared_ptr<IOKitEventContext>;
 using IOKitSubscriptionContextRef = std::shared_ptr<IOKitSubscriptionContext>;
