+++ osquery/tables/system/darwin/kernel_info.cpp
@@ -16,7 +16,7 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
-#include "osquery/tables/system/darwin/iokit_utils.h"
+#include "osquery/events/darwin/iokit.h"
 #include "osquery/tables/system/efi_misc.h"
 
 namespace osquery {
