+++ osquery/tables/system/darwin/sip_config.cpp
@@ -13,7 +13,7 @@
 #include <osquery/sql.h>
 #include <osquery/logger.h>
 
-#include "osquery/tables/system/darwin/iokit_utils.h"
+#include "osquery/events/darwin/iokit.h"
 
 namespace osquery {
 namespace tables {
