+++ osquery/tables/system/darwin/smbios_tables.cpp
@@ -18,8 +18,8 @@
 
 #include <osquery/tables.h>
 
+#include "osquery/events/darwin/iokit.h"
 #include "osquery/tables/system/smbios_utils.h"
-#include "osquery/tables/system/darwin/iokit_utils.h"
 
 namespace osquery {
 namespace tables {
